refactor: move card settings to directory
